Update token vnode for api tokens with infinite timeouts
The former implementaiton of the vnode was enforcing timeouts
on all tokens, that does not work any more with api tokens. The
new vnode keeps two databases, ones for timed out tokens and one
for tokens w/o a timeout. This might slow things down just a little
bit but it's required to prevent unused tokens to leak and start
filling the disk.

In the process there were a few bugs where the dict was used to
access 'all' tokens instead of the database - this is fixed.

Since we're R18+ now dict is fully repalced by maps.