Improve tokens, add api and revoke for tokens, cleanp
Moved the scopes to maps from tuples, it's cleaner. Added functions
for creating API Keys and revoking tokens from a TokenID including
the related API calls. the OAuth Application context is now a
map too instead of a record.

Some old functions are removed that were directly accessing the
user elements instead of using the general getter and ft_* calls.